requirements.txt
fastapi
uvicorn[standard]
boto3
langgraph
langchain-aws
chromadb>=0.5.0
sqlalchemy
pydantic
mcp
python-dotenv
langgraph-checkpoint-sqlite
arize
arize-otel
openinference-instrumentation-langchain
tavily-python
pandas
openpyxl
